+++ lib/compiler/npm_package.rb
@@ -44,16 +44,21 @@ class Compiler
       else
         raise Error, "No binaries detected inside #{@package_path}."
       end
-      if @binaries[@bin_name]
+      if @binaries.kind_of?(Hash) && @binaries[@bin_name]
         STDERR.puts "Using #{@bin_name} at #{@binaries[@bin_name]}"
+        bin = @binaries[@bin_name]
+      elsif @binaries.kind_of?(String)
+        STDERR.puts "\n\nWARNING: Ignored supplied entrance `#{@bin_name}` since `bin` of package.json is a string\n\n"
+        STDERR.puts "Using entrance #{@binaries}"
+        bin = @binaries
       else
         raise Error, "No such binary: #{@bin_name}"
       end
-      ret = File.expand_path("node_modules/#{@module_name}/#{@binaries[@bin_name]}")
+      ret = File.expand_path("node_modules/#{@module_name}/#{bin}")
       unless File.exist?(ret)
         raise Error, "Npm install failed to generate #{ret}"
       end
-      return File.expand_path("node_modules/#{@module_name}/#{@binaries[@bin_name]}", @work_dir)
+      return File.expand_path("node_modules/#{@module_name}/#{bin}", @work_dir)
     end
   end
 end